Split Unity callstacks with both System.Environment.NewLine and \n separators
Unity 5.5 on Windows generates internal callstacks which are separated with \n, not \r\n. This is probably because Unity has internal string-concatenation code which happens to be hardcoded to use \n on all platforms. This change makes the Unity internal callstack splitting code first try with the OS version (either \n or \r\n depending on OS) and, if the split operation generated an identical string, tries again with \n hardcoded. This makes exception callstacks decode correctly on Unity 5.5/Windows.