Structured file logger overwrite fixes + more
When logging to file, you can now choose whether to overwrite the existing file or to create a new unique file instead. This enables you to launch the same executable multiple times from the same folder and getting a unique log file for each application instance.

StructuredFileLogger is initialized in Awake(). This puts more of  the startup activity into the log file. For best results, the user will also need to change the script execution order such that UberLoggerLogToFile is initialized first.