Advertising some UberLogger add-ons
I would like to make visible some of the extensions that I have written. UberLogger-StructuredFile in particular is something that I think would be beneficial to many others.

How about adding them to a section of the readme like this? If not, is there some other reasonable way that others can discover the extensions from the UberLogger core repo?